Senior Project/back-end/app/requirements.txt
absl-py==2.1.0
aio-pika==9.5.5
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiormq==6.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.1
deepeval==2.3.8
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gevent==24.11.1
geventhttpclient==2.3.3
google-ai-generativelanguage==0.6.6
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gotrue==2.11.3
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.9.0
kiwisolver==1.4.8
lancedb==0.16.0
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.35
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.8
libclang==18.1.1
llama-cloud==0.1.12
llama-cloud-services==0.6.1
llama-index==0.12.18
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.18
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.7
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
locust==2.33.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
openai==1.63.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pamqp==3.3.0
pandas==2.2.3
pika==1.3.2
pillow==10.3.0
pluggy==1.5.0
portalocker==3.1.1
postgrest==0.19.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.1
Pygments==2.18.0
pylance==0.19.2
pyparsing==3.2.0
pypdf==5.3.0
pytest==8.3.4
pytest-asyncio==0.26.0
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.3.0
realtime==2.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.41.3
storage3==0.11.3
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0